% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Cleaning.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean Earthquake data}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{Raw data frame with columns YEAR, MONTH, DAY, LATITUDE and LONGITUDE}
}
\value{
Data frame with additional column DATE
}
\description{
Clean up the Earthquake Dataframe
}
\details{
This function cleans up the data frame in the sense that it creates a new DATE Column and converting LATITUDE and LONGITUDE columns to numeric
}
\examples{
eq_clean_data(df)

}
