% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maps.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Map Annotations}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{dataframe containing the columns LOCATION_NAME, EQ_PRIMARY and DEATHS}
}
\value{
Label Character vector in html format
}
\description{
Create Annotation Column for use in eq_map
}
\details{
This function creates a text annotation column in html
with the format VariableName: Value for Location, Magnitude and Total Deaths of each earthquake in seperate rows.
Where the data is missing the corresponding row will be omitted. Can be used for example with the dplyr::mutate function (see examples).
}
\examples{
\dontrun{
readr::read_delim("results.tsv", delim = "\\t") \%>\% eq_clean_data() \%>\%
 dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\% eq_map(annot_col = "popup_text")
}
}
