% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Cleaning.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Earthquake Location}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{Raw data frame with column LOCATION_NAME}
}
\value{
Data frame with transformed LOCATION_NAME column
}
\description{
Clean up the Earthquake Location column
}
\details{
This function removes the country name from the location by splitting the location column at the first colon
}
\examples{
eq_clean_location(df)

}
