% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maps.R
\name{eq_map}
\alias{eq_map}
\title{Earthquake Map}
\usage{
eq_map(df, annot_col = "DATE")
}
\arguments{
\item{df}{Raw data frame with columns LONGITUDE, LATITUDE, EQ_PRIMARY}

\item{annot_col}{Column name to use as text for annotations}
}
\value{
Interactive map object as defined by the leaflet package
}
\description{
Create interactive html-Earthquake map based on leaflet package
}
\details{
This function creas an interactive html map with Earthquake Locations marked with Circles whose
radius is proportional to the Earthquake magnitude (EQ_PRIMARY). The text supplied in the annotation column will be displayed upon clicking
on a specific Earthquake marker
}
\examples{
\dontrun{
readr::read_delim("results.tsv", delim = "\\t") \%>\% eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\% eq_map(annot_col = "DATE")
}
}
