% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geoms.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{data}{The data to be displayed in this layer. There are three
   options:

   If `NULL`, the default, the data is inherited from the plot
   data as specified in the call to [ggplot()].

   A `data.frame`, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   [fortify()] for which variables will be created.

   A `function` will be called with a single argument,
   the plot data. The return value must be a `data.frame.`, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`color = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
`NA`, the default, includes if any aesthetics are mapped.
`FALSE` never includes, and `TRUE` always includes.}

\item{inherit.aes}{If `FALSE`, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. [borders()].}

\item{geom}{The geometric object to use display the data}

\item{check.aes, check.param}{If `TRUE`, the default, will check that
supplied parameters and aesthetics are understood by the `geom` or
`stat`. Use `FALSE` to suppress the checks.}

\item{params}{Additional parameters to the `geom` and `stat`.}

\item{subset}{DEPRECATED. An older way of subsetting the dataset used in a
layer.}
}
\description{
Timeline Visualisation of Earthquake data
}
\details{
The timeline geom visualises Earthquakes as points on one or multiple parallel timelines.
The size of the points as well as the colour can be used to visualize additional variables.
}
\section{Aesthetics}{

\code{geom_timeline} understands the following asthetics:
\itemize{
 \item{"x"}{ Variable (normally date) positioning the earthquake on the timeline}
 \item{"y} { Variable for plotting data on multiple parallel timelines}
 \item{"size"}{ Radius of the points on the timeline}
 \item{"shape"}{ Shape of the points on the timeline}
 \item{"Colour"}{ Colour of the points on the timeline}
}
}

\examples{
ggplot(data = df[COUNTRY \%in\% c("CHINA","USA") & YEAR >= 2000], aes(x = DATE, y = COUNTRY, color = DEATHS, size = EQ_PRIMARY)) + geom_timeline(alpha = 0.2)
}
